// lib/api/generated/models/CatalogMediaAssetRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CatalogMediaAssetRead = {
    id: string;
    provider?: string;
    provider_file_id?: (string | null);
    name?: (string | null);
    url: string;
    thumbnail_url?: (string | null);
    file_path?: (string | null);
    mime_type?: (string | null);
    size_bytes?: (number | null);
    width?: (number | null);
    height?: (number | null);
    alt_text?: (string | null);
    metadata?: Record<string, any>;
    created_at?: (string | null);
    updated_at?: (string | null);
    created_by?: (string | null);
};